0.71-Auto-generate protos
These Powershell scripts will generate new proto files based on the SRC folder structure and data. It updates all new types, enums etc, so that everything is up to date, based on the newest 0.x.x.proto file. It also creates a proto_missing.txt file to show which files are found in original .proto file, but has not been generated. As a safety measure, inline Enums are copied anyway. This is just to make sure that no information is missing.

!Don't worry about running this! It exports all new file to a NEW folder so that it is easy to compare the results and you can still decide what to use or not.

#Instructions to use.

First generate the index files to read the source folders to identify all source import etc. Start with the command:

> split_discover_protos.bat

Then when the index is complete, you can run the generation command. This will create all new files in the TEMP_OUT folder, so you can do a comparison with the original. Also it wil identify all missing types and files, and stores these in the file proto_missing.txt

Just run

> split_proto_run.bat

#Please note:

Although this is an automated process, I cannot guarantee a 100% correct generation of the files. So please still check the output and update the "split_proto.ps1" with new fixes.